backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
mysql-connector-python==9.4.0
pydantic==2.11.7
python-dotenv==1.0.1
python-multipart==0.0.6
feedparser==6.0.11
requests==2.32.3
markdownify==0.11.6
tencentcloud-sdk-python==3.0.1155